Finish invoice implementation. Show lable with sats on dashboard chat list

+++ sphinx/application/common/wrappers/wrapper-message/src/main/java/chat/sphinx/wrapper_message/Msg.kt
@@ -8,7 +8,7 @@ import java.lang.IllegalArgumentException
 @JsonClass(generateAdapter = true)
 data class Msg(
     val content: String?,
-    val amount: Int?,
+    val amount: Long?,
     val mediaToken: String?,
     val mediaKey: String?,
     val mediaType: String?,
